Moved CountdownLabelActivity to separate module.
