+++ app/src/main/res/values/strings.xml
@@ -101,25 +101,4 @@
 
     <string name="next">Next</string>
     <string name="loading">Loading…</string>
-
-    <plurals name="days">
-        <item quantity="zero">%d Days</item>
-        <item quantity="one">%d Day</item>
-        <item quantity="other">%d Days</item>
-    </plurals>
-    <plurals name="hours">
-        <item quantity="zero">%d Hours</item>
-        <item quantity="one">%d Hour</item>
-        <item quantity="other">%d Hours</item>
-    </plurals>
-    <plurals name="minutes">
-        <item quantity="zero">%d Minutes</item>
-        <item quantity="one">%d Minute</item>
-        <item quantity="other">%d Minutes</item>
-    </plurals>
-    <plurals name="seconds">
-        <item quantity="zero">%d Seconds</item>
-        <item quantity="one">%d Second</item>
-        <item quantity="other">%d Seconds</item>
-    </plurals>
 </resources>
